requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# WhatsApp Cloud API
pywa==4.0.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Cache / queue
redis==5.2.1
celery==5.4.0

# Config
pydantic-settings==2.7.0
python-dotenv==1.0.1

# HTTP client
httpx==0.28.1

# Ingestion (Module 4 — multi-source mandi price pipeline)
beautifulsoup4==4.12.3
lxml==5.3.0
tenacity==9.0.0

# LLM fallback
google-generativeai==0.8.3

# Speech-to-Text (Phase 2 Module 2 — voice messages)
google-cloud-speech==2.23.0
# openai library (for Whisper fallback, optional)

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
